% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loom_and_analyze.R
\name{loom_too_Seurat}
\alias{loom_too_Seurat}
\title{Read and process loom and H5AD files from FlyCellAtlas}
\usage{
loom_too_Seurat(
  loom_file = "./loom/s_fca_biohub_body_wall_10x.loom",
  h5ad_file = "./loom/s_fca_biohub_body_wall_10x.h5ad",
  plot = T
)
}
\arguments{
\item{loom_file}{character with Name and directory of loom file -h5da has to be in same directory-}

\item{h5ad_file}{character with Name and directory of h5da file}

\item{plot}{Boolean indicating if plots should be generated}

\item{dir}{Directory to save RData and plots in}
}
\value{
Seurat object from paper without any further processing
}
\description{
This function requires loom and h5ad files from FlyCellAtlas and process them into Seurat objects.
}
\examples{
There are examples in the folder names "./loom". To run it:
loom_to_seurat(loom_file = "~/Documents/scripts/FlyCellAtlas_download_with_annotations/loom/s_fca_biohub_body_wall_10x.loom")

}
