% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loom_and_analyze.R
\name{loom_too_Seurat_brandNew}
\alias{loom_too_Seurat_brandNew}
\title{Read and process loom and H5AD files from FlyCellAtlas}
\usage{
loom_too_Seurat_brandNew(
  loom_file = "./loom/s_fca_biohub_body_wall_10x.loom",
  h5ad_file = "./loom/s_fca_biohub_body_wall_10x.h5ad",
  plot = T,
  preprocesing = T,
  dir = "./"
)
}
\arguments{
\item{loom_file}{character with Name and directory of loom file -h5da has to be in same directory-}

\item{h5ad_file}{character with Name and directory of h5da file}

\item{plot}{Boolean indicating if plots should be generated}

\item{preprocesing}{Boolean indicating if preprocessing should be done (ie. findVariableFeatues, Scale Data, etc) or not}

\item{dir}{Directory to save plots in}
}
\value{
It saves old and new Seurat onjects along with marker genes and metadata.
*"_h5" is seurat object from the paper, this only has most variable genes

*"_new" is the new seurat object

*"_new_markers" is all.marker genes from the new seurat object with all genes

*"_paper_markers" is all.marker genes from the paper

*,"_matadata" is the metadata
}
\description{
This function requires loom and h5ad files from FlyCellAtlas and process them into Seurat objects. It creates a new object form scratch and tranfere celltype assignment
}
\examples{
There are examples in the folder names "./loom". To run it:
loom_too_Seurat_brandNew(loom_file = "~/Documents/scripts/FlyCellAtlas_download_with_annotations/loom/s_fca_biohub_body_wall_10x.loom")

}
